Fix GCC/Linux build portability issues
Several MSVC-tolerated patterns broke the build under GCC 13 on Linux,
found while verifying a from-scratch Linux build:

- Missing <algorithm>/<cstring>/<cmath>/<stdexcept> includes relied on
  transitive inclusion that GCC's libstdc++ doesn't provide (ChunkWater,
  Model, Misc.h, ZoneIDBrowser, minimap_widget).
- std::exception has no (const char*) constructor outside MSVC; switched
  the affected throw sites to std::runtime_error (Tool, MapView,
  area_trigger, AreaTriggerEditor).
- WorldRenderParams::editing_mode/display_mode members shared their
  enum's name, which GCC rejects as "changes meaning of" without an
  elaborated enum specifier.
- WMOInstance::draw had a duplicated `const` on two reference parameters.
- NoggitWindow used a unique_ptr to a forward-declared type with no
  out-of-line destructor, so libstdc++ instantiated default_delete
  against an incomplete type from an unrelated moc translation unit.
- GroundEffectsTool passed float* into an overload of modf that only
  matches double* once unqualified sin()/modf() resolve to the global
  C functions.
- A few Qt includes only worked on case-insensitive filesystems
  (QCheckBox.h/QComboBox.h/QLabel.h/QSpinBox.h, QtImGui vs qtimgui,
  qtemporaryfile) and fail on ext4/btrfs.

Verified with a full Release build against Qt 5.15 / GCC 13.

+++ src/noggit/ui/GroundEffectsTool.cpp
@@ -1041,7 +1041,7 @@ namespace Noggit
             {
                 // Stable by record id so the legend, newly loaded tiles and
                 // undo/redo-driven renderer refreshes all agree.
-                float partr, partg, partb;
+                double partr, partg, partb;
                 float const id = static_cast<float>(effect.ID);
                 float r = std::abs(modf(sin(id * 12.9898f + 78.233f) * 43758.5453f, &partr));
                 float g = std::abs(modf(sin(id * 11.5591f + 70.233f) * 43569.5451f, &partg));
